Refactoring + ça marche sauf le md5 check.

+++ bit_alterne/server.h
@@ -1,7 +1,7 @@
 #ifndef SERVER_H
 #define SERVER_H
 
-int send_ack(int sockfd, struct sockaddr *addr, uint32_t seq);
+int send_ack(int sockfd, struct sockaddr *addr, uint8_t seq);
 int main(int argc, char *argv[]);
